--- Sources/DeclarativeUIKit/Environment/UIKitEnvironmentKey.swift
@@ -1,7 +1,0 @@
-
-import UIKit
-
-public protocol UIKitEnvironmentKey {
-    associatedtype Value
-    static var defaultValue: Self.Value { get }
-}
